+++ conftest.py
@@ -5,5 +5,5 @@ from synch.factory import init
 
 @pytest.fixture(scope="session", autouse=True)
 def initialize_tests():
-    # init('tests/synch.ini')
+    # init('tests/synch.yaml')
     pass