+++ synch/reader/__init__.py
@@ -2,7 +2,7 @@ import abc
 import logging
 import signal
 from signal import Signals
-from typing import Callable, Tuple, Union, Dict
+from typing import Callable, Dict, Tuple, Union
 
 from synch.broker import Broker
 
@@ -15,10 +15,10 @@ class Reader:
 
     def __init__(self, source_db: Dict):
         self.source_db = source_db
-        self.host = source_db.get('host')
-        self.port = source_db.get('port')
-        self.user = source_db.get('user')
-        self.password = source_db.get('password')
+        self.host = source_db.get("host")
+        self.port = source_db.get("port")
+        self.user = source_db.get("user")
+        self.password = source_db.get("password")
 
         signal.signal(signal.SIGINT, self.signal_handler)
         signal.signal(signal.SIGTERM, self.signal_handler)