+++ synch/replication/etl.py
@@ -1,7 +1,7 @@
 import logging
 from typing import Dict
 
-from synch import get_writer, Global
+from synch import get_writer
 from synch.enums import ClickHouseEngine, SourceDatabase
 from synch.settings import Settings
 
@@ -13,11 +13,11 @@ def get_source_select_sql(schema: str, table: str, source_db: Dict, addition_col
     get source db select sql from different db
     """
     select = "*"
-    user = source_db.get('user')
-    host = source_db.get('host')
-    port = source_db.get('port')
-    password = source_db.get('password')
-    db_type = source_db.get('db_type')
+    user = source_db.get("user")
+    host = source_db.get("host")
+    port = source_db.get("port")
+    password = source_db.get("password")
+    db_type = source_db.get("db_type")
     if addition_column:
         select += f", toInt8(1) as {addition_column}"
     if db_type == SourceDatabase.postgres.value:
@@ -26,8 +26,16 @@ def get_source_select_sql(schema: str, table: str, source_db: Dict, addition_col
         return f"SELECT {select} FROM mysql('{host}:{port}', '{schema}', '{table}', '{user}', '{password}')"
 
 
-def get_table_create_sql(schema: str, table: str, pk: str, partition_by: str, engine_settings: str,
-                         engine: ClickHouseEngine, source_db: Dict, **kwargs):
+def get_table_create_sql(
+    schema: str,
+    table: str,
+    pk: str,
+    partition_by: str,
+    engine_settings: str,
+    engine: ClickHouseEngine,
+    source_db: Dict,
+    **kwargs,
+):
     """
     get table create sql from by settings
     """
@@ -41,16 +49,14 @@ def get_table_create_sql(schema: str, table: str, pk: str, partition_by: str, en
         select_sql = get_source_select_sql(schema, table, source_db)
         return f"CREATE TABLE {schema}.{table} ENGINE = {engine} {partition_by_str} ORDER BY {pk} {engine_settings_str} AS {select_sql} limit 0"
     elif engine == ClickHouseEngine.collapsing_merge_tree:
-        sign_column = kwargs.get('sign_column')
+        sign_column = kwargs.get("sign_column")
         select_sql = get_source_select_sql(schema, table, source_db, sign_column)
         return f"CREATE TABLE {schema}.{table} ENGINE = {engine}({sign_column}) {partition_by_str} ORDER BY {pk} {engine_settings_str} AS {select_sql} limit 0"
 
 
-def get_full_insert_sql(
-        schema: str, source_db: Dict, source_db_database_table: Dict
-):
+def get_full_insert_sql(schema: str, source_db: Dict, source_db_database_table: Dict):
     engine = source_db_database_table.get("clickhouse_engine")
-    table = source_db_database_table.get('table')
+    table = source_db_database_table.get("table")
     if engine == ClickHouseEngine.merge_tree:
         return f"insert into {schema}.{table} {get_source_select_sql(schema, table, source_db)}"
     elif engine == ClickHouseEngine.collapsing_merge_tree:
@@ -59,18 +65,20 @@ def get_full_insert_sql(
 
 
 def etl_full(
-        reader, settings: Settings, schema: str, tables_pk: Dict, alias: str, renew=False,
+    reader, settings: Settings, schema: str, tables_pk: Dict, alias: str, renew=False,
 ):
     """
     full etl
     """
     source_db = settings.get_source_db(alias)
     source_db_database = settings.get_source_db_database(alias, schema)
-    schema = source_db_database.get('database')
-    for table in source_db_database.get('tables'):
-        table_name = table.get('table')
+    schema = source_db_database.get("database")
+    for table in source_db_database.get("tables"):
+        if not table.get("auto_full_etl"):
+            continue
+        table_name = table.get("table")
         pk = tables_pk.get(table_name)
-        writer = get_writer(table.get('clickhouse_engine'))
+        writer = get_writer(table.get("clickhouse_engine"))
         if not pk:
             logger.warning(f"No pk found in {schema}.{table_name}, skip")
             continue
@@ -85,59 +93,20 @@ def etl_full(
                 logger.warning(f"Try to drop table {schema}.{table_name} fail")
         if not writer.table_exists(schema, table_name):
             writer.execute(
-                get_table_create_sql(schema, table_name, pk, table.get('partition_by'), table.get('engine_settings'),
-                                     table.get('clickhouse_engine'), source_db))
+                get_table_create_sql(
+                    schema,
+                    table_name,
+                    pk,
+                    table.get("partition_by"),
+                    table.get("engine_settings"),
+                    table.get("clickhouse_engine"),
+                    source_db,
+                )
+            )
             if reader.fix_column_type:
                 writer.fix_table_column_type(reader, schema, table_name)
-            source_db_database_table = settings.get_source_db_database_table(alias, schema, table_name)
+            source_db_database_table = settings.get_source_db_database_table(
+                alias, schema, table_name
+            )
             writer.execute(get_full_insert_sql(schema, source_db, source_db_database_table))
-            logger.info(f"etl success:{schema}.{table_name}")
-
-
-def finish_continuous_etl():
-    logger.info("finish success, bye!")
-    Global.broker.close()
-    exit()
-
-
-def handle_event_by_engine():
-    """
-    handle event by different engine
-    :return:
-    """
-    pass
-
-
-def continuous_etl(schema: str, tables_pk: Dict, last_msg_id, skip_error: bool, insert_interval: int):
-    """
-    continuous etl from broker and insert into clickhouse
-    :param insert_interval:
-    :param schema:
-    :param tables_pk:
-    :param last_msg_id:
-    :param skip_error:
-    :return:
-    """
-    len_event = 0
-    is_stop = False
-    broker = Global.broker
-    for msg_id, msg in broker.msgs(
-            schema, last_msg_id=last_msg_id, block=insert_interval * 1000
-    ):
-        if not msg_id:
-            if len_event > 0:
-                is_insert = True
-                alter_table = False
-                query = None
-            else:
-                if is_stop:
-                    finish_continuous_etl()
-                continue
-        else:
-            logger.debug(f"msg_id:{msg_id}, msg:{msg}")
-            len_event += 1
-            event = msg
-            table = event["table"]
-            schema = event["schema"]
-            action = event["action"]
-
+            logger.info(f"full data etl for {schema}.{table_name} success")