+++ synch/writer/__init__.py
@@ -21,13 +21,11 @@ class ClickHouse:
     def __init__(self, settings: Dict):
         self.settings = settings
         self._client = clickhouse_driver.Client(
-            host=settings.get('host'),
-            port=settings.get('port'),
-            user=settings.get('user'),
-            password=settings.get('password') or '',
+            host=settings.get("host"),
+            port=settings.get("port"),
+            user=settings.get("user"),
+            password=settings.get("password") or "",
         )
-        signal.signal(signal.SIGINT, self.signal_handler)
-        signal.signal(signal.SIGTERM, self.signal_handler)
 
     def table_exists(self, schema: str, table: str):
         sql = f"select count(*)from system.tables where database = '{schema}' and name = '{table}'"
@@ -64,57 +62,9 @@ class ClickHouse:
                 )
             self.execute(fix_sql)
 
-    def check_mutations(self, schema: str):
-        """
-        检查mutations是否有失败的(ch后台异步的update和delete变更)
-        """
-        mutation_list = ["mutation_failed", "table", "create_time"]
-        fail_list = []
-        mutation_data = []
-        query_sql = (
-            f"select count(*) from system.mutations where is_done=0 and database = '{schema}'"
-        )
-        mutation_sql = f"select count(*) as mutation_faild ,concat(database,'.',table)as db,create_time from system.mutations where is_done=0 and database = '{schema}' group by db,create_time"
-        mutations_failed_num = self.execute(query_sql)[0][0]
-        if mutations_failed_num >= 10:
-            fail_data = self.execute(mutation_sql)
-            for d in fail_data:
-                fail_list.append(list(d))
-            for d in fail_list:
-                tmp = dict(zip(mutation_list, d))
-                mutation_data.append(tmp)
-            last_data = json.dumps(mutation_data, indent=4, cls=JsonEncoder)
-            message = "mutations error failed num {0}. delete error please check: {1}".format(
-                mutations_failed_num, last_data
-            )
-            logger.error(message)
-
     def insert_events(self, schema: str, table: str, insert_data: List[Dict]):
         insert_sql = "insert into {0}.{1} values ".format(schema, table)
         self.execute(insert_sql, list(map(lambda x: x.get("values"), insert_data)))
-
-    def signal_handler(self, signum: Signals, handler: Callable):
-        sig = Signals(signum)
-        if self.len_event == 0:
-            logger.info(f"shutdown consumer on {sig.name} success")
-            exit()
-        else:
-            logger.info(
-                f"shutdown consumer on {sig.name}, wait {self.settings.insert_interval} seconds at most to finish {self.len_event} events..."
-            )
-            self.is_stop = True
-            self.is_insert = True
-
-
-
-    def after_insert(self, schema: str, events_num: int):
-        self.broker.commit(schema)
-        logger.info(f"success commit {events_num} events")
-        self.event_list = {}
-        self.is_insert = False
-        self.len_event = 0
-        if self.is_stop:
-            self.finish()
 
     def alter_table(self, query: str, skip_error: bool):
         """
@@ -129,7 +79,3 @@ class ClickHouse:
         else:
             self.execute(query)
             logger.info(f"execute ddl query: {query}")
-
-    @abc.abstractmethod
-    def start_consume(self, schema: str, tables_pk: Dict, last_msg_id, skip_error: bool):
-        raise NotImplementedError